Dropdown menu pour choisir Dubins ou Bezier
